refactor: move threading to infrastructure layer
- Move threading.rs to infrastructure/threading/simulation_runner.rs
- Update all imports across the codebase
- Threading logic now properly in infrastructure layer
- Following clean architecture separation of concerns
