+++ src/infrastructure/threading/simulation_runner.rs
