+++ src/simulation/mod.rs
@@ -1,7 +1,6 @@
 pub mod entities;
 pub mod map;
 pub mod robot_ai;
-pub mod threading;
 pub mod visualization;
 
 #[cfg(test)]
