+++ src/simulation/robot_ai/executor.rs
@@ -2,7 +2,7 @@ use crate::simulation::entities::{Map, Station};
 use crate::simulation::robot_ai::pathfinding::Pathfinder;
 use crate::simulation::robot_ai::robot::Robot;
 use crate::simulation::robot_ai::types::{AnalyzeTask, ExploreTask, HarvestTask, RobotState};
-use crate::simulation::threading::SimulationMessage;
+use crate::infrastructure::threading::simulation_runner::SimulationMessage;
 use tokio::sync::mpsc::Sender;
 
 pub struct Executor;
