+++ src/simulation/robot_ai/robot.rs
@@ -4,7 +4,7 @@ use crate::simulation::robot_ai::behavior::RobotBehavior;
 use crate::simulation::robot_ai::executor::Executor;
 use crate::simulation::robot_ai::pathfinding::Pathfinder;
 use crate::simulation::robot_ai::types::{Direction, RobotState, RobotType, Task, TaskType};
-use crate::simulation::threading::SimulationMessage;
+use crate::infrastructure::threading::simulation_runner::SimulationMessage;
 use tokio::sync::mpsc::Sender;
 
 pub struct Robot {
